+++ explcheck/src/explcheck-issues.lua
@@ -1,5 +1,6 @@
 -- A registry of warnings and errors identified by different processing steps.
 
+local format_identifier = require('explcheck-format').format_issue_identifier
 local get_option = require("explcheck-config").get_option
 
 local Issues = {}
@@ -10,17 +11,22 @@ function Issues.new(cls, pathname, options)
   setmetatable(self, cls)
   cls.__index = cls
   -- Initialize the class.
+  self.closed = false
+  --- Issue tables
   self.errors = {}
   self.warnings = {}
+  --- Seen issues
   self.seen_issues = {}
+  --- Suppressed issues
   self.suppressed_issue_map = {}
   for issue_identifier, suppressed_issues in pairs(get_option("suppressed_issue_map", options, pathname)) do
     issue_identifier = self._normalize_identifier(issue_identifier)
     self.suppressed_issue_map[issue_identifier] = suppressed_issues
   end
+  --- Ignored issues
   self.ignored_issues = {}
   for _, issue_identifier in ipairs(get_option("ignored_issues", options, pathname)) do
-    self:ignore(issue_identifier)
+    self:ignore({identifier_prefix = issue_identifier})
   end
   return self
 end
@@ -45,6 +51,10 @@ end
 
 -- Add an issue to the table of issues.
 function Issues:add(identifier, message, range, context)
+  if self.closed then
+    error('Cannot add issues to a closed issue registry')
+  end
+
   identifier = self._normalize_identifier(identifier)
 
   -- Discard duplicate issues.
@@ -53,11 +63,14 @@ function Issues:add(identifier, message, range, context)
   if self.seen_issues[identifier] == nil then
     self.seen_issues[identifier] = {}
   end
-  if self.seen_issues[identifier][range_start] == nil then
-    self.seen_issues[identifier][range_start] = {}
+  if self.seen_issues[identifier][context or ''] == nil then
+    self.seen_issues[identifier][context or ''] = {}
   end
-  if self.seen_issues[identifier][range_start][range_end] == nil then
-    self.seen_issues[identifier][range_start][range_end] = true
+  if self.seen_issues[identifier][context or ''][range_start] == nil then
+    self.seen_issues[identifier][context or ''][range_start] = {}
+  end
+  if self.seen_issues[identifier][context or ''][range_start][range_end] == nil then
+    self.seen_issues[identifier][context or ''][range_start][range_end] = true
   else
     return
   end
@@ -66,7 +79,7 @@ function Issues:add(identifier, message, range, context)
   if self.suppressed_issue_map[identifier] ~= nil then
     for _, suppressed_issue_identifier in ipairs(self.suppressed_issue_map[identifier]) do
       suppressed_issue_identifier = self._normalize_identifier(suppressed_issue_identifier)
-      self:ignore(suppressed_issue_identifier, range)
+      self:ignore({identifier_prefix = suppressed_issue_identifier, range = range, seen = true})
     end
   end
 
@@ -74,8 +87,9 @@ function Issues:add(identifier, message, range, context)
   local issue = {identifier, message, range, context}
 
   -- Determine if the issue should be ignored.
-  for _, ignore_issue in ipairs(self.ignored_issues) do
-    if ignore_issue(issue) then
+  for _, ignored_issue in ipairs(self.ignored_issues) do
+    if ignored_issue.check(issue) then
+      ignored_issue.seen = true
       return
     end
   end
@@ -86,33 +100,37 @@ function Issues:add(identifier, message, range, context)
 end
 
 -- Prevent issues from being present in the table of issues.
-function Issues:ignore(identifier_prefix, range)
-  if identifier_prefix ~= nil then
-    identifier_prefix = self._normalize_identifier(identifier_prefix)
+function Issues:ignore(ignored_issue)
+  if self.closed then
+    error('Cannot ignore issues in a closed issue registry')
+  end
+
+  if ignored_issue.identifier_prefix ~= nil then
+    ignored_issue.identifier_prefix = self._normalize_identifier(ignored_issue.identifier_prefix)
   end
 
   -- Determine which issues should be ignored.
   local function match_issue_range(issue_range)
     return (
-      issue_range:start() >= range:start() and issue_range:start() <= range:stop()  -- issue starts within range
-      or issue_range:start() <= range:start() and issue_range:stop() >= range:stop()  -- issue is in middle of range
-      or issue_range:stop() >= range:start() and issue_range:stop() <= range:stop()  -- issue ends within range
+      issue_range:start() >= ignored_issue.range:start() and issue_range:start() <= ignored_issue.range:stop()  -- issue starts within range
+      or issue_range:start() <= ignored_issue.range:start() and issue_range:stop() >= ignored_issue.range:stop() -- issue in middle of range
+      or issue_range:stop() >= ignored_issue.range:start() and issue_range:stop() <= ignored_issue.range:stop()  -- issue ends within range
     )
   end
   local function match_issue_identifier(issue_identifier)
     -- Match the prefix of an issue, allowing us to ignore whole sets of issues with prefixes like "s" or "w4".
-    return issue_identifier:sub(1, #identifier_prefix) == identifier_prefix
+    return issue_identifier:sub(1, #ignored_issue.identifier_prefix) == ignored_issue.identifier_prefix
   end
 
-  local ignore_issue, issue_tables
-  if identifier_prefix == nil and range == nil then
+  local issue_tables
+  if ignored_issue.identifier_prefix == nil and ignored_issue.range == nil then
     -- Prevent any issues.
     issue_tables = {self.warnings, self.errors}
-    ignore_issue = function() return true end
-  elseif identifier_prefix == nil then
+    ignored_issue.check = function() return true end
+  elseif ignored_issue.identifier_prefix == nil then
     -- Prevent any issues within the given range.
     issue_tables = {self.warnings, self.errors}
-    ignore_issue = function(issue)
+    ignored_issue.check = function(issue)
       local issue_range = issue[3]
       if issue_range == nil then  -- file-wide issue
         return false
@@ -120,34 +138,44 @@ function Issues:ignore(identifier_prefix, range)
         return match_issue_range(issue_range)
       end
     end
-  elseif range == nil then
+  elseif ignored_issue.range == nil then
     -- Prevent any issues with the given identifier.
-    assert(identifier_prefix ~= nil)
-    issue_tables = {self:_get_issue_table(identifier_prefix)}
-    ignore_issue = function(issue)
+    assert(ignored_issue.identifier_prefix ~= nil)
+    issue_tables = {self:_get_issue_table(ignored_issue.identifier_prefix)}
+    ignored_issue.check = function(issue)
       local issue_identifier = issue[1]
       return match_issue_identifier(issue_identifier)
     end
   else
     -- Prevent any issues with the given identifier that are also either within the given range or file-wide.
-    assert(range ~= nil and identifier_prefix ~= nil)
-    issue_tables = {self:_get_issue_table(identifier_prefix)}
-    ignore_issue = function(issue)
+    assert(ignored_issue.range ~= nil and ignored_issue.identifier_prefix ~= nil)
+    issue_tables = {self:_get_issue_table(ignored_issue.identifier_prefix)}
+    ignored_issue.check = function(issue)
       local issue_identifier = issue[1]
       local issue_range = issue[3]
       if issue_range == nil then  -- file-wide issue
-        return match_issue_identifier(issue_identifier)
+        -- Make an exception for issue S105 (Needlessly ignored issue), which can be both ranged and file-wide.
+        -- Only match the ranged variant of this issue here.
+        if issue_identifier == self._normalize_identifier('s105') then
+          return false
+        else
+          return match_issue_identifier(issue_identifier)
+        end
       else  -- ranged issue
         return match_issue_range(issue_range) and match_issue_identifier(issue_identifier)
       end
     end
   end
+  assert(ignored_issue.check ~= nil)
+  assert(issue_tables ~= nil)
 
   -- Remove the issue if it has already been added.
   for _, issue_table in ipairs(issue_tables) do
     local filtered_issues = {}
     for _, issue in ipairs(issue_table) do
-      if not ignore_issue(issue) then
+      if ignored_issue.check(issue) then
+        ignored_issue.seen = true
+      else
         table.insert(filtered_issues, issue)
       end
     end
@@ -160,11 +188,15 @@ function Issues:ignore(identifier_prefix, range)
   end
 
   -- Prevent the issue from being added later.
-  table.insert(self.ignored_issues, ignore_issue)
+  table.insert(self.ignored_issues, ignored_issue)
 end
 
 -- Check whether two registries only contain issues with the same codes.
 function Issues:has_same_codes_as(other)
+  if not self.closed or not other.close then
+    error('Cannot compared issues between unclosed issue registries')
+  end
+
   -- Collect codes of all issues.
   local self_codes, other_codes = {}, {}
   for _, table_name in ipairs({'warnings', 'errors'}) do
@@ -189,6 +221,27 @@ function Issues:has_same_codes_as(other)
     end
   end
   return true
+end
+
+-- Close the issue registry, preventing future modifications and report all needlessly ignored issues.
+function Issues:close()
+  if self.closed then
+    error('Cannot close an already closed issue registry')
+  end
+
+  -- Report all needlessly ignored issues.
+  for _, ignored_issue in ipairs(self.ignored_issues) do
+    if not ignored_issue.seen and ignored_issue.identifier_prefix ~= self._normalize_identifier('s105') then
+      local formatted_identifier_prefix
+      if ignored_issue.identifier_prefix ~= nil then
+        formatted_identifier_prefix = format_identifier(ignored_issue.identifier_prefix)
+      end
+      self:add('s105', 'Needlessly ignored issue', ignored_issue.source_range, formatted_identifier_prefix)
+    end
+  end
+
+  -- Close the registry.
+  self.closed = true
 end
 
 -- Sort the warnings/errors using location as the primary key.